requirements.txt
django
python-dotenv
django-bootstrap4
django-icons
django-debug-toolbar
django-sslserver
djangorestframework
